+++ CHANGES.md
@@ -36,6 +36,7 @@
 * Change py2 print statements to py2/3 compatible functions
 * Change Kodi notifier to use requests as opposed to urllib
 * Change to consolidate scene exceptions and name cache code
+* Add BeyondHD provider
 
 [develop changelog]
 * Update Requests library 2.7.0 (ab1f493) to 2.7.0 (8b5e457)